Exercício Vetor

+++ Ex02Vetores.java
@@ -9,7 +9,7 @@ public class Ex02Vetores {
 		Scanner leia = new Scanner (System.in);
 		
 		int x, soma = 0;
-		double media;
+		float media;
 		int vetor[] = new int[10];
 		
 		for(x=0 ; x<10 ; x++) {
